# Add a missing GC preserve in Ryu