This PR replaces a use of `pointer` with `Base.cconvert`, `GC.@preserve`, and `Base.unsafe_convert`.