@oscardssmith 2025-10-04
Does this really need to use unsafe at all?