@nhz2 2025-10-04
Here is a basic benchmark:

```julia
using BenchmarkTools
buf = zeros(UInt8, 10000);
@btime Base.Ryu.writeshortest($buf, 1, 1234.123456)
```

v1.11.7:  23.633 ns
v1.12.0-rc3: 23.180 ns
master: 23.312 ns
This PR: 24.725 ns

So using `Base.unsafe_copyto!` seems slightly slower than `memmove` here.