@vtjnash 2025-10-04
yes, there is an extra boundscheck there potentially, especially without an `@inbounds`, but it seems generally fine